services/gradebook-service/requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose[cryptography]==3.3.0
pytest==8.3.4
httpx==0.27.2

